Former-commit-id: 8c879b1eb8fd272359378ec3ee3ab0adafa1ee10

+++ Sources/Internal/Render/Highlevel/RenderSystem.cpp
@@ -81,7 +81,7 @@ void RenderSystem::RenderPermanent(RenderObject* renderObject)
     /*on add calculate valid world bbox*/
     renderObject->Retain();
     renderObjectArray.push_back(renderObject);
-    renderObject->SetRemoveIndex((uint32)(renderObjectArray.size() - 1));
+    renderObject->SetRemoveIndex(static_cast<uint32>(renderObjectArray.size() - 1));
 
     AddRenderObject(renderObject);
     //    uint32 renderBatchCount = renderObject->GetRenderBatchCount();
@@ -298,8 +298,8 @@ void RenderSystem::FindNearestLights(RenderObject* renderObject)
 
 void RenderSystem::FindNearestLights()
 {
-    uint32 size = (uint32)renderObjectArray.size();
-    for (uint32 k = 0; k < size; ++k)
+    size_t size = renderObjectArray.size();
+    for (size_t k = 0; k < size; ++k)
     {
         FindNearestLights(renderObjectArray[k]);
     }
